# swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "雨诺供应链协同平台Restful接口定义\n"
  version: "1.0.0"
  title: "Swagger SccApi"
host: "127.0.0.1:3300"
basePath: "/api/v1"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /authentication:
    post:
      tags:
      - "authentication"
      summary: "登录认证接口"
      description: "\"根据用户名&密码去验证用户,确认密码与用户名无误后,将会返回用户的基本信息: operatorInfo\\n\"\n"
      operationId: "postAuthentication"
      parameters:
      - name: "device-type"
        in: "header"
        description: "表示用户的设备类型,用来做一些优化使用\n"
        required: false
        type: "string"
        enum:
        - "android"
        - "react-pc"
        - "react-phone"
      - in: "body"
        name: "userInfo"
        description: "登录用户名"
        required: true
        schema:
          $ref: "#/definitions/authenticationInfo"
      responses:
        default:
          description: "响应对象feedback"
          schema:
            $ref: "#/definitions/getAuthenticationFeedback"
      x-swagger-router-controller: "Authentication"
...
definitions:
  getAuthenticationFeedback:
    type: "object"
    required:
    - "msg"
    - "status"
    properties:
      status:
        type: "integer"
        format: "int32"
        description: "feedback status: |\n  请求处理最后的状态\n"
      msg:
        type: "string"
        description: "对用户友好的提示信息\n"
      data:
        $ref: "#/definitions/loginResData"
...
  loginResData:
    type: "object"
    properties:
      nextTo:
        type: "string"
        description: "保留字段"
        maxLength: 50
      token:
        type: "string"
        description: "token信息"
        maxLength: 50
      operaterInfo:
        $ref: "#/definitions/loginResData_operaterInfo"
    description: "data对象中存放了响应的数据"
  authenticationInfo:
    type: "object"
    properties:
      username:
        type: "string"
        description: "用户的登录名"
      password:
        type: "string"
        description: "用户密码"
...
  loginResData_operaterInfo:
    properties:
      clientId:
        type: "integer"
        description: "客户ID"
      failCount:
        type: "integer"
        description: "登录失败计数器"
      clientName:
        type: "string"
        description: "客户名称"
      bannedTime:
        type: "string"
        description: "时间"
      operatorRoles:
        type: "string"
        description: "操作员权限描述"
      customerId:
        type: "integer"
        description: "商户ID"
      operatorType:
        type: "string"
        description: "操作员类型 枚举值"
        enum:
        - "CUSTOMER"
        - "CLIENT"
      operatorId:
        type: "integer"
        description: "操作员ID"
      operatorName:
        type: "string"
        description: "操作员名称"
      username:
        type: "string"
        description: "操作员名称"
...